requirements.txt
flask
flask-smorest
python-dotenv
flask-pymongo
flask-jwt-extended
passlib